Initialize $fullquery_string and update empty check
Previously, the variable $fullquery_string was not initialized,
leading to incorrect behavior with recent versions of PHP due to a
change in the behavior of count(NULL). This commit correctly
initializes the variable and adds !is_null($fullquery_string) to
count($fullquery_string) for enhanced robustness.

(internal: 6fdc64a, 1ce2fce)